+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.53
+*INCOMPLETE*
+
 ## 0.0.52
 + Added `beta.OpenVINOInferenceMetadata` class to lower PyTorch models for inference with Intel OpenVINO.
 + Added `beta.QnnInferenceMetadata` class to lower PyTorch models for inference on Qualcomm accelerators with QNN SDK.